+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # WorkerID ChangeLog
 
-## Unreleased
+## WorkerID v0.3.1
 
 ### Bug fixes